Consolidated all rectangle logic into a parent class GameObject

+++ src/main/java/arcadeGame/Wall.java
@@ -22,7 +22,7 @@ public class Wall extends Tile {
 	public void display(Graphics2D g) {
 		g.translate(getX(), getY());
 		g.setColor(Color.gray);
-		g.fillRect(0, 0, width, height);
+		g.fillRect(0, 0, (int) width, (int) height);
 		g.translate(-getX(), -getY());
 	}
